"Fixed bugs and minor modifications in UCLutils
Modified priorMask3D and Fit2Mua2Mus to allow TOAST mapping without preliminary mapping to DOTgrid (upon uncommenting of lines)

Modified Fit2Mua2Mus to allow reconstruction of absorption only (in and out of inclusion)

"

+++ src/UCLutils/pointSplineSegs.m
@@ -13,7 +13,7 @@ function [maskIm] = pointSplineSegs(import)
  % think of using dicomread(imName) with dicom images...to be checked
 
 disp('Segmentation...');
-maskIm = roispline(import);
+maskIm = roispline(uint8(import));
 
 
 figure;
